Chore: Rename src files from .js to .jsx
This commit renames all relevant JavaScript files within the 'src'
directory (including components, pages, tests, and entry points like
index.js and App.js) to use the .jsx extension.

This change is made to explicitly denote files containing JSX syntax,
which is a common convention in React projects.

All corresponding relative import statements throughout the project
have been updated to point to the new .jsx file extensions.
