+++ react-admin-dashboard/src/App.test.jsx
@@ -1,10 +1,10 @@
 import { render, screen } from '@testing-library/react';
-import App from './App';
+import App from './App.jsx';
 
 // Mock child components to isolate App.js tests and avoid deep rendering issues
-jest.mock('./components/Sidebar', () => () => <aside data-testid="sidebar">Mock Sidebar</aside>);
-jest.mock('./components/Header', () => () => <header data-testid="header">Mock Header</header>);
-jest.mock('./components/MainContent', () => () => <main data-testid="main-content">Mock MainContent</main>);
+jest.mock('./components/Sidebar.jsx', () => () => <aside data-testid="sidebar">Mock Sidebar</aside>);
+jest.mock('./components/Header.jsx', () => () => <header data-testid="header">Mock Header</header>);
+jest.mock('./components/MainContent.jsx', () => () => <main data-testid="main-content">Mock MainContent</main>);
 
 describe('App Component', () => {
   test('renders without crashing', () => {
